position, rotation, move, and rotate support
- Entities can be specifically transformed by position or rotation
- Added @#position = vec3, @#rotation = mat3, @#move(vec3), and @#rotate(mat3)

+++ src/render_support/SpritePhysics.cpp
@@ -109,7 +109,7 @@ void SpritePhysics::update()
 {
     mTimeCurr = now();
     f64 delta = mTimeCurr - mTimePrev;
-    mpDynamicsWorld->stepSimulation(delta, 4);
+    mpDynamicsWorld->stepSimulation((f32)delta, 4);
     mTimePrev = mTimeCurr;
 
     // Check for collisions
